requirements.txt
streamlit
openai
langchain
pinecone
python-dotenv
langchain-community
langchain-pinecone
tiktoken